pyproject.toml
[project]
name = "agentic-tv-controller"
version = "0.1.0"
description = "Control Android TV with natural language via LangChain"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn>=0.34.0",
    "langchain-openai>=0.3.0",
    "langchain-core>=0.3.0",
    "python-dotenv>=1.0.0",
    "asyncpg>=0.30.0",
    "pydantic-settings>=2.0.0",
    "pillow>=10.0.0",
    "pytesseract>=0.3.10",
]

[project.scripts]
agentic-tv = "server:main"
